rationalize the tinymce config
9 files changed, 85 insertions(+), 432 deletions(-)
